requirements.txt
psycopg2-binary==2.9.7
pandas==1.5.3
numpy==1.24.3
elasticsearch==8.7.0
jupyterlab==4.0.5
python-dotenv==1.0.0